Use maxsplits to only examine a single event at a time
Previously the code split buf on all end_of_field matches, and then
recombined any remaning unprocessed event with "".join.  This
caused multiple newlines that may have matched to be lost.
We only want to deal with one event at a time in __next__,
so only the part of buf up to the first end_of_field
separator should be touched.  The rest of buf should
be left alone until next time.

Fixes #5